Finished course without concurrency

+++ tasks/intro/multiplication/multiplication.cpp
@@ -3,5 +3,5 @@
 #include <stdexcept>
 
 int64_t Multiply(int a, int b) {
-    return static_cast<int64_t>(a) * static_cast<int64_t>(b);
+    return static_cast<int64_t>(a) * b;
 }
